% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microplot_cover}
\alias{microplot_cover}
\title{Herbaceous plant cover from two microplots in each of five subplots at 24 sites at the UC Sedgwick Reserve.}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3946 rows and 5 columns.}
\usage{
microplot_cover
}
\description{
A dataframe for average plant cover from two microplots per each of the 
five subplots at each of the 24 sites at Sedgwick the UC Sedgwick Reserve in 
Santa Barbara Co. CA, USA. Data were collected in the spring of 2017 by 
visually estimating percent cover in two 10 x 10 cm microplots within each
25 x 25 cm subplot at each site. Cover estimates were made in mid-spring
and late spring and the maximum cover per species per site is reported.
}
\keyword{datasets}
