% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_cover_2019}
\alias{plot_cover_2019}
\title{Spring 2019 plant cover from sampling plots at 24 sites at Sedgwick}
\format{A data frame with 10920 rows and 8 variables:
\describe{
  \item{year}{year of cover estimate}
  \item{site}{site number, 740 to 763}
  \item{plot}{plot number, 1 to 5}
  \item{area}{plot area in cm^2}
  \item{USDA_symbol}{USDA symbol used to identify species}
  \item{cover}{estimated cover of each species in each plot}
  \item{count_type}{Indicates whether counts are direct counts of individuals across the entire plot or estimated counts made by counting individuals in 15 x 15 cm subplot and extroplating that density to the entire plot.}
  \item{count}{number of individuals of each species in the plot}
}}
\usage{
plot_cover_2019
}
\description{
A dataframe of cover estimates and total number of individuals of all vascular plants 
found in 55 x 55 cm sampling plots at the UC Sedgwick Reserve. There are five plots at each 
of the 24 study sites (n = 120). Individual counts are either direct counts across the 
entire plot area or are extrapolated to the whole plot area by counting all the individuals 
in a representative 15 x 15 cm subplot within the plot. Data collected by Andy Kleinhesselink 
in April of 2019.
}
\keyword{datasets}
