% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_cover_2017}
\alias{site_cover_2017}
\title{Spring 2017 average plant cover for 24 study sites at Sedgwick}
\format{A data frame with 2016 rows and 7 variables:
\describe{
  \item{year}{year of cover estimate}
  \item{site}{site number, 740 to 763}
  \item{area}{Total sample area: 5 plots x 625 cm^2 per site}
  \item{USDA_symbol}{USDA symbol used to identify species}
  \item{n_plots}{number of plots where the species was present in each site}
  \item{sd_cover}{standard deviation of cover across the plots at each site}
  \item{avg_cover}{average cover of species across the five plots at each site}
}}
\usage{
site_cover_2017
}
\description{
A dataframe of average cover for all vascular plant species found at each of the 24 
study sites at the UC Sedgwick Reserve. Cover is an average of across five 
25 x 25 cm sampling plots located at each site. Data collected 
by Andy Kleinhesselink in April of 2017.'
}
\keyword{datasets}
