% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_cover_2019}
\alias{site_cover_2019}
\title{Spring 2019 average plant cover and density at 24 sites at Sedgwick}
\format{A data frame with 2184 rows and 9 variables:
\describe{
  \item{year}{year of cover estimate}
  \item{site}{site number, 740 to 763}
  \item{area}{total sampled area: 5 plots x 3025 cm^2 per site}
  \item{USDA_symbol}{USDA symbol used to identify each species}
  \item{n_plots}{number of plots where the species was present in each site}
  \item{sd_cover}{standard deviation of cover across the plots at each site}
  \item{avg_cover}{average cover of species across the five plots at each site}
  \item{tot_count}{total number of individuals of each species across the five plots at each site}
  \item{sd_count}{standard deviation of counts across the five plots at each site}
}}
\usage{
site_cover_2019
}
\description{
A dataframe of average cover and total individual abundance for all vascular plants 
found at each of the 24 study sites at the UC Sedgwick Reserve. Cover is averaged and total 
count is summed across across five 55 x 55 cm sampling plots located at each site. Data collected 
by Andy Kleinhesselink in April of 2019.'
}
\keyword{datasets}
