% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subplot_cover}
\alias{subplot_cover}
\title{Herbaceous plant cover from five subplots at 24 sites at the UC Sedgwick Reserve.}
\format{A data frame with 1973 rows and 3 variables:
\describe{
  \item{site}{Site number 740 to 763} 
  \item{subplot}{ Subplot per site, numbered 1 to 5} 
  \item{species}{latin binomial of species, using calflora taxonomy}
  \item{cover}{cover of each species}
}}
\usage{
subplot_cover
}
\description{
A dataframe for plant cover from five subplots at each of 24 sites at
the UC Sedgwick Reserve in Santa Barbara Co. CA, USA. Data were collected in 
the spring of 2017 by visually estimating percent cover in each of 5 25 x 25 cm 
subplots at each site. Cover estimates were made in mid-spring
and late spring and the maximum cover per species per site is reported.
}
\keyword{datasets}
