% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_cover_2017}
\alias{plot_cover_2017}
\title{Spring 2017 plant cover from sampling plots at each of the 24 sites at Sedgwick.}
\format{A data frame with 10080 rows and 8 variables:
\describe{
  \item{year}{year of cover estimate}
  \item{site}{site number, 740 to 763}
  \item{plot}{plot number, 1 to 5}
  \item{USDA_symbol}{USDA symbol used to identify species}
  \item{area}{Sample area in cm^2}
  \item{cover}{estimated cover of each species in each plot}
}}
\usage{
plot_cover_2017
}
\description{
A dataframe of cover estimates for all vascular plant species found in five
20 x 20 cm plots located along a 10 m transect at each of the 24 study sites at 
the UC Sedgwick Reserve. Data collected by Andy Kleinhesselink in April of 2017.
}
\keyword{datasets}
