% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_cover_2017}
\alias{plot_cover_2017}
\title{Herbaceous plant cover from five subplots at 24 sites at the UC Sedgwick Reserve.}
\format{A data frame with 1973 rows and 6 variables:
\describe{
  \item{year}{year of cover estimate}
  \item{site}{site number, 740 to 763}
  \item{plot}{plot number, 1 to 5}
  \item{USDA_symbol}{USDA symbol used to identify species}
  \item{area_cm2}{Total sample area in cm^2}
  \item{cover}{cover of each species per plot}
}}
\usage{
plot_cover_2017
}
\description{
A dataframe for plant cover from five subplots at each of 24 sites at
the UC Sedgwick Reserve in Santa Barbara Co. CA, USA. Data were collected in 
the spring of 2017 by visually estimating percent cover in each of 5 25 x 25 cm 
subplots at each site. Cover estimates were made in mid-spring
and late spring and the maximum cover per species per site is reported.
}
\keyword{datasets}
