% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_cover_2017}
\alias{site_cover_2017}
\title{Herbaceous plant cover from 24 sites at the UC Sedgwick Reserve.}
\format{A data frame with 397 rows and 5 variables:
\describe{
  \item{year}{year of cover estimate}
  \item{site}{site number, 740 to 763}
  \item{USDA_symbol}{USDA symbol used to identify species}
  \item{area_cm2}{Total sample area in cm^2}
  \item{cover}{average cover of each species across the five subplots at each site}
}}
\usage{
site_cover_2017
}
\description{
A data frame for average plant cover from 24 monitoring sites at 
the UC Sedgwick Reserve in Santa Barbara Co. CA, USA. Data were collected in 
the spring of 2017 by visually estimating percent cover in 5 25 x 25 cm subplots
Cover is average cover across all plots. Cover estimates were made in mid-spring
and late spring and the maximum cover per species per site was used.
}
\keyword{datasets}
