% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subplot_cover_2017}
\alias{subplot_cover_2017}
\title{Spring 2017 plant cover from microplots at Sedgwick.}
\format{A data frame with 20170 rows and 7 variables:
\describe{
  \item{year}{year of cover estimate}
  \item{site}{Site number 740 to 763.}
  \item{plot}{plot number, 1 to 5}
  \item{subplot}{Subplot within each subplot, identified as LL and UR, lower left and upper right respectively.}
  \item{area}{Total sample area in cm^2}
  \item{USDA_symbol}{USDA symbol used to identify species}
  \item{cover}{cover of each species per plot}
}}
\usage{
subplot_cover_2017
}
\description{
A dataframe of cover estimates for all vascular plant species found in 
10 x 10 cm microplots nested within each plot at each of the 24 sampling sites
at the UC Sedgwick Reserve. Data collected by Andy Kleinhesselink in April of 2017.
}
\keyword{datasets}
